requirements.txt
# Twitch VOD to YouTube Archiver
# Install with: pip install -r requirements.txt

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP requests for Twitch API
requests>=2.31.0

# Video downloading
yt-dlp>=2024.1.1


